Refactor monolithic test_backbone.py into focused per-module test files
Split the single 2090-line test_backbone.py into 9 smaller test files
organized by source module, following NetworkX algorithm test conventions:

- tests/conftest.py: shared graph-builder fixtures
- tests/test_structural.py: global threshold, strongest-N ties, salience,
  metric/ultrametric, MST, doubly stochastic, h-backbone, modularity, PMFG
- tests/test_neighborhood.py: overlap, Jaccard, Dice, cosine backbones
- tests/test_statistical.py: disparity, noise-corrected, marginal
  likelihood, ECM, LANS filters
- tests/test_hybrid.py: GLAB filter
- tests/test_bipartite.py: SDSM, FDSM
- tests/test_unweighted.py: sparsify, LSpar, local degree
- tests/test_filters.py: threshold, fraction, boolean, consensus
- tests/test_measures.py: node/edge/weight fraction, reachability, KS
- tests/test_integration.py: end-to-end pipelines, edge cases, smoke tests

Converted class-based tests to standalone test functions and extracted
reusable graph builders into pytest fixtures in conftest.py. All 212
tests pass.
